{{cookiecutter.repo_name}}/requirements/prod.txt
# python web framework
# django==1.10.8
django==1.11.16

# django restful api
# djangorestframework==3.7.7
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
django-filter==2.0.0
Werkzeug==0.14.1
django-guardian==1.4.9

# database driver
pymysql==0.9.2
mysqlclient==1.3.13

# markdown
markdown==3.0.1


# shell
ipython==7.1.1
httpie==1.0.2

# image
Pillow==5.3.0

# virtual env
python-dotenv==0.9.1

# task and message queue
celery[redis]==4.2.1
django_celery_beat==1.3.0
django_celery_results==1.0.4

# other
django-mptt==0.9.1
